implement stacked bar chart and progressBar
